Ikkoku: split Program
