+++ Ikkoku/src/CommandLine/CheckCmd.cs
@@ -1,10 +1,32 @@
 ﻿using Mobsub.SubtitleParse.AssTypes;
 using Mobsub.Ikkoku.SubtileProcess;
+using System.CommandLine;
+using System.IO;
 
 namespace Mobsub.Ikkoku.CommandLine;
 
 internal class CheckCmd
 {
+    internal static Command Build(Argument<FileSystemInfo> path, Option<bool> verbose)
+    {
+        var tagMode = new Option<string>(
+            name: "--tag",
+            description: "Check tags mode: mod, weird, both.").FromAmong("mod", "weird", "both");
+        var styleCheck = new Option<bool>(
+            name: "--style",
+            description: "Check undefined styles.");
+
+        var checkCommand = new Command("check", "Check Your ASS!")
+        {
+            path, tagMode, styleCheck, verbose
+        };
+        checkCommand.SetHandler(Execute, path, tagMode, styleCheck, verbose);
+
+        // fonts glyphs subcommand
+
+        return checkCommand;
+    }
+
     internal static void Execute(FileSystemInfo path, string? tagMode, bool styleCheck, bool verbose)
     {
         switch (path)
